requirements.txt
ffmpeg-python==0.2.0
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
filelock==3.0.12
future==0.17.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.17
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.4
packaging==19.0
pluggy==0.12.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==4.6.1
pytest-mock==1.10.4
pytz==2019.1
requests==2.22.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
toml==0.10.0
tox==3.12.1
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
zipp==0.5.1
